Add test for segment based evaluation
